feat: api client fixes, more tests, better linter (#3)
* feat: add docs, remove unnecessary jobs list, better swagger codegen script

* add more docs and tests

* a bit more docs

* test: spec the utils

* fmt the gen api client

* fmt the api

* fmt main.go

* docs: update README and add make scripts for test cluster

* ci: rework gitlab

* test: test converters, utils

* test: more converters tests

* test: fix some tests in converters

* test: reformat printer

* feat: change namespace and deployment targets to store dynamic data in status subres

* Revert "feat: change namespace and deployment targets to store dynamic data in status subres"

This reverts commit 9d5fb47499fcc6d1f6bf2de8055287be91ca8fef.

* refactor: make the deployment fetching smoother

* feat: update the deployment targets client to have 201 response

* feat: better makefile test clustering

+++ hack/update-swagger-codegen.sh
@@ -26,13 +26,13 @@ SWAGGER_IGNORE_FILE=.swagger-codegen-ignore
 CONFIG_FILE=swagger-gen-config.json
 OUT_DIR=ververica-platform-api
 
-rm -rf ${PROJECT_DIR}/${OUT_DIR}
-mkdir ${PROJECT_DIR}/${OUT_DIR}
-cp ${PROJECT_DIR}/${SWAGGER_IGNORE_FILE} ${PROJECT_DIR}/${OUT_DIR}
+rm -rf "${PROJECT_DIR:?}"/${OUT_DIR}
+mkdir "${PROJECT_DIR}"/${OUT_DIR}
+cp "${PROJECT_DIR}"/${SWAGGER_IGNORE_FILE} "${PROJECT_DIR}"/${OUT_DIR}
 
 docker run --rm \
-    -v ${PROJECT_DIR}:/local:rw \
-    --user $(id -u):$(id -g) \
+    -v "${PROJECT_DIR}":/local:rw \
+    --user "$(id -u)":"$(id -g)" \
     swaggerapi/swagger-codegen-cli generate \
     -i /local/${SWAGGER_FILE} \
     -l go \
